Work in progress.. Tree browser now seems to work, even on a VT100 terminal. Some heavy renaming done. Probably not yet a consistent build. DO NOT USE.

+++ software/userinterface/context_menu.h
@@ -21,7 +21,7 @@ class ContextMenu : public UIObject
     bool perform_quick_seek(void);
 
 	// private functions:
-    virtual int handle_key(BYTE c);
+    virtual int handle_key(int c);
 
     Contextable *contextable;
